[PM] Disable more of the loop passes -- LCSSA and LoopSimplify are also not really wired into the loop pass manager in a way that will let us productively use these passes yet.
This lets the new PM get farther in basic testing which is useful for
establishing a good baseline of "doesn't explode". There are still
plenty of crashers in basic testing though, this just gets rid of some
noise that is well understood and not representing a specific or narrow
bug.